chore: refactoring and improve aria
