+++ packages/wired-icon/src/iconset/icon-set-communication.ts
