+++ packages/wired-icon/src/iconset/icon-set-content.ts
