+++ packages/wired-icon/src/iconset/icon-set-device.ts
