+++ packages/wired-icon/src/iconset/icon-set-editor.ts
