+++ packages/wired-icon/src/iconset/icon-set-hardware.ts
