+++ packages/wired-icon/src/iconset/icon-set-image.ts
