+++ packages/wired-icon/src/iconset/icon-set-maps.ts
