+++ packages/wired-icon/src/iconset/icon-set-places.ts
